+++ src/main/java/com/sipgate/li/lib/x1/protocol/error/GenericBaseException.java
@@ -7,7 +7,7 @@ abstract class GenericBaseException extends ErrorResponseException {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(GenericBaseException.class);
 
-  protected GenericBaseException(final Integer errorCode, final Exception e) {
+  protected GenericBaseException(final long errorCode, final Exception e) {
     // don't expose possibly confidential information from generic exception
     super(errorCode);
     // but write everything to the log so developers can debug
